// Task Cockpit/src/Workspace/Scopes/Definition.d.ts
import type * as TC from '../../types';

/** "Описание задачи" (не сама задача), извлечённое из
 * разобранного определения задачи в файле.
 *
 * Нормализованное представление пользовательских полей
 * (с валидацией и значениями по умолчанию), минимально необходимое
 * для представления задачи в UI. */
interface Definition {

    /** Флаг скрытия из отображения */
    hidden: boolean;

    /** Пользовательская иконка */
    icon: TC.Icon | null;

    /** Признак фонового выполнения. */
    isBackground: boolean;

    /** Группа выполнения. */
    group: TC.TaskGroup | null;

    /** Имя */
    name: TC.TaskName;

}


export default Definition;